docs(tree): update Tree.d.ts and Tree Docs
