+++ src/TreePicker/TreePicker.d.ts
@@ -2,30 +2,9 @@ import * as React from 'react';
 
 import { FormControlPickerProps } from '../@types/common';
 import { TreeBaseProps } from '../Tree/TreeBase';
+import { TreeDragProps } from '../Tree/TreeDrag';
 
-/**
- * Tree Node Drag Type
- */
-export enum TREE_NODE_DROP_POSITION {
-  DRAG_OVER = 0, // drag node in tree node
-  DRAG_OVER_TOP = 1, // drag node on tree node
-  DRAG_OVER_BOTTOM = 2 // drag node under tree node
-}
-
-export interface DropData {
-  /** drag node data */
-  dragNode: any;
-
-  /** dropNode data */
-  dropNode: any;
-
-  /** node drop postion */
-  dropNodePosition: TREE_NODE_DROP_POSITION;
-
-  /** Update Data when drop node */
-  createUpdateDataFunction: (data: any[]) => any[];
-}
-export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
+export interface TreePickerProps extends TreeBaseProps, TreeDragProps, FormControlPickerProps {
   /** The height of Dropdown */
   height?: number;
 
@@ -35,14 +14,11 @@ export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
   /** A picker that can be counted */
   countable?: boolean;
 
-  /** Whether dispaly search input box */
+  /** Whether display search input box */
   searchable?: boolean;
 
   /** Whether using virtualized list */
   virtualized?: boolean;
-
-  /** Whether the node can  be dragged */
-  draggable?: boolean;
 
   /** Set the option value for the expand node */
   defaultExpandItemValues?: any[];
@@ -62,26 +38,6 @@ export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
 
   /** Called when scrolling */
   onScroll?: (event: React.SyntheticEvent<HTMLElement>) => void;
-
-  /** Called when node drag start */
-  onDragStart?: (nodeData: any, e: React.DragEvent) => void;
-
-  /** Called when node drag enter */
-  onDragEnter?: (nodeData: any, e: React.DragEvent) => void;
-
-  /** Called when node drag over */
-  onDragOver?: (nodeData: any, e: React.DragEvent) => void;
-
-  /** Called when node drag leave */
-  onDragLeave?: (nodeData: any, e: React.DragEvent) => void;
-
-  /** Called when node drag end */
-  onDragEnd?: (nodeData: any, e: React.DragEvent) => void;
-
-  /** Called when node drop */
-  onDrop?: (dropData: DropData, e: React.DragEvent) => void;
-
-  renderDragNode?: (dragNode: any) => React.ReactNode;
 }
 
 declare const TreePicker: React.ComponentType<TreePickerProps>;
